Abstract Item link field to ChecklistLinkField view
